fix route admin

+++ routes/web.php
@@ -42,33 +42,33 @@ Route::prefix('admin')->middleware(['auth', 'admin'])->group(function () {
         Route::get('/', [DashboardController::class, 'index'])->name('admin.dashboard');
     });
     Route::prefix('category')->group(function () {
-        Route::get('/', [AdminCategoryController::class, 'index'])->name('category.index');
-        Route::post('/search', [AdminCategoryController::class, 'search'])->name('category.search');
-        Route::post('/create', [AdminCategoryController::class, 'create'])->name('category.create');
-        Route::post('/update', [AdminCategoryController::class, 'update'])->name('category.update');
-        Route::delete('/delete/{id}', [AdminCategoryController::class, 'delete'])->name('category.delete');
+        Route::get('/', [AdminCategoryController::class, 'index'])->name('admin.category.index');
+        Route::post('/search', [AdminCategoryController::class, 'search'])->name('admin.category.search');
+        Route::post('/create', [AdminCategoryController::class, 'create'])->name('admin.category.create');
+        Route::post('/update', [AdminCategoryController::class, 'update'])->name('admin.category.update');
+        Route::delete('/delete/{id}', [AdminCategoryController::class, 'delete'])->name('admin.category.delete');
     });
 
     Route::prefix('products')->group(function () {
-        Route::get('/', [AdminProductController::class, 'index'])->name('product.index');
-        Route::post('/search', [AdminProductController::class, 'search'])->name('product.search');
-        Route::post('/create', [AdminProductController::class, 'create'])->name('product.create');
-        Route::post('/update', [AdminProductController::class, 'update'])->name('product.update');
-        Route::delete('/delete/{id}', [AdminProductController::class, 'delete'])->name('product.delete');
+        Route::get('/', [AdminProductController::class, 'index'])->name('admin.product.index');
+        Route::post('/search', [AdminProductController::class, 'search'])->name('admin.product.search');
+        Route::post('/create', [AdminProductController::class, 'create'])->name('admin.product.create');
+        Route::post('/update', [AdminProductController::class, 'update'])->name('admin.product.update');
+        Route::delete('/delete/{id}', [AdminProductController::class, 'delete'])->name('admin.product.delete');
     });
 
     Route::prefix('posts')->group(function () {
-        Route::get('/', [AdminPostController::class, 'index'])->name('post.index');
-        Route::post('/search', [AdminPostController::class, 'search'])->name('post.search');
-        Route::post('/create', [AdminPostController::class, 'create'])->name('post.create');
-        Route::post('/update', [AdminPostController::class, 'update'])->name('post.update');
-        Route::delete('/delete/{id}', [AdminPostController::class, 'delete'])->name('post.delete');
+        Route::get('/', [AdminPostController::class, 'index'])->name('admin.post.index');
+        Route::post('/search', [AdminPostController::class, 'search'])->name('admin.post.search');
+        Route::post('/create', [AdminPostController::class, 'create'])->name('admin.post.create');
+        Route::post('/update', [AdminPostController::class, 'update'])->name('admin.post.update');
+        Route::delete('/delete/{id}', [AdminPostController::class, 'delete'])->name('admin.post.delete');
     });
 
     Route::prefix('contacts')->group(function () {
-        Route::get('/', [AdminContactController::class, 'index'])->name('contact.index');
-        Route::post('/search', [AdminContactController::class, 'search'])->name('contact.search');
-        Route::delete('/delete/{id}', [AdminContactController::class, 'delete'])->name('contact.delete');
+        Route::get('/', [AdminContactController::class, 'index'])->name('admin.contact.index');
+        Route::post('/search', [AdminContactController::class, 'search'])->name('admin.contact.search');
+        Route::delete('/delete/{id}', [AdminContactController::class, 'delete'])->name('admin.contact.delete');
     });
     Route::post('ckeditor/upload', [CKEditorController::class, 'upload'])->name('ckeditor.upload');
 });
